requirements.txt
bcrypt==4.2.1
cffi==1.17.1
colorama==0.4.6
cryptography==44.0.0
mysql-connector-python==9.2.0
paramiko==3.5.0
pycparser==2.22
PyNaCl==1.5.0
pyspnego==0.11.2
smbprotocol==1.15.0
telnetlib-313-and-up==3.13.1; python_version >= '3.13'
